refactor : 입력 관련 함수 모듈화
- '구현 목록: 입력' 관련 모듈 따로 분리

this refers to #9
