+++ src/index.js
@@ -1,59 +1,23 @@
 import RacingCarGame from './racingCarGame.js';
-import ValidationError from './validationError.js';
+import Scanner from './scanner.js';
 
 const racingCarGame = new RacingCarGame();
-
-const getCarNamesToList = function () {
-  const carNamesInput = document.querySelector('#car-names-input');
-  const regexp = /[^,\s]+[^,]*[^,\s]+/gi; // ,와 양쪽 공백을 제외한 부분만 가져온다
-  const carNamesList = carNamesInput.value.match(regexp);
-
-  if (!carNamesList) {
-    throw new ValidationError('자동차 이름을 입력해주세요', carNamesInput);
-  }
-  if (carNamesList.some(arr => arr.length > 5)) {
-    throw new ValidationError(
-      '자동차 이름은 5자 이하로 입력해주세요',
-      carNamesInput,
-    );
-  }
-
-  return carNamesList;
-};
-
-const setElemVisible = function (elem, visible) {
-  if (visible) {
-    elem.classList.replace('hide', 'visible');
-  } else {
-    elem.classList.replace('visible', 'hide');
-  }
-};
 
 // TODO: 이름 다시 세팅하면 게임 결과 지우기
 const handleCarNamesSubmitBtn = function () {
   try {
-    const carNamesList = getCarNamesToList();
+    const carNamesList = Scanner.getCarNamesInputToList();
     racingCarGame.initCars(carNamesList);
-    setElemVisible(document.querySelector('#racing-count-container'), true);
+    Scanner.setElemVisible(
+      document.querySelector('#racing-count-container'),
+      true,
+    );
   } catch (e) {
     alert(e.message);
     if (e.errElem) {
       e.errElem.focus();
     }
   }
-};
-
-const getRacingCount = function () {
-  const racingCountInput = document.querySelector('#racing-count-input');
-  const racingCount = Number(racingCountInput.value);
-  if (!racingCountInput.value || Number.isNaN(racingCount)) {
-    throw new ValidationError('숫자를 입력해주세요', racingCountInput);
-  }
-  if (!Number.isInteger(racingCount) || racingCount <= 0) {
-    throw new ValidationError('자연수를 입력해주세요', racingCountInput);
-  }
-
-  return racingCount;
 };
 
 const printRoundResult = function (printContainer) {
@@ -82,7 +46,7 @@ const printWinner = function (printContainer) {
 
 const handleCountSubmitBtn = function () {
   try {
-    const racingCount = getRacingCount();
+    const racingCount = Scanner.getRacingCountFromUser();
     const carGameResultContainer = document.querySelector(
       '#car-game-result-container',
     );
@@ -94,7 +58,7 @@ const handleCountSubmitBtn = function () {
       printRoundResult(carGameResultContainer);
     }
     printWinner(carGameResultContainer);
-    setElemVisible(carGameResultContainer, true);
+    Scanner.setElemVisible(carGameResultContainer, true);
   } catch (e) {
     // TODO: 알림 메소드 따로 빼기
     alert(e.message);
